link3D works
tweaked stuff to compile for matlab

fixed bug in matlab_process

bunch of other changes

changed MATLAB_COMPILE to MATLAB_MEX_FILE


git-svn-id: svn://innoue/tracking/trunk@101 4cb6fdf0-3dc9-4801-8ffc-c6659000aa44

+++ common_src/cxx/corr_goftheta.cpp
@@ -156,7 +156,7 @@ void Corr_goftheta::compute(const particle * p_in,const vector<const particle*> 
       
       // assume the nhood is the right size
       
-      int ind = floor(n_bins_ * (theta)/(pi));
+      int ind = (int)floor(n_bins_ * (theta)/(pi));
       if(ind ==n_bins_)
 	ind = 0;
       try
